Update volcanoes.json
It produced about 104–207 km3 (25–50 cu mi) of tephra (several times as much as the 1980 eruption of Mount St. Helens), produced major pyroclastic fall and large pyroclastic flows that covered 10,000 km2 (3,900 sq mi) with over 50 cm (20 in) of pumice and ash, and nearly 2,000,000 km2 (770,000 sq mi) with a blanket of ash at least 0.5 cm (0.20 in) in depth. The eruption devastated an area of up to 100 km (62 mi) radius around the volcano.